Baymax is human friendly health companion app which helps user to identify any kind of disease user might have.
Baymax is a simple health companion andriod app which provides a user various kinds of diseases which a user might have. It needs a user input as what are the symptoms the user is experiencing. The app analyzes the user input and provides the list of diseases according to the symptoms the user possibily might have. The diseases list will be displayed priority wise i.e. most probable disease that user might have will be on first position and likewise the other diseases, if there, will be shown on the screen. It's not complete yet, our team is working on it. Hold on till it successfully runs and make people's life more easy.